requirements.txt
requests
google-genai
python-dotenv
flask
fastapi
uvicorn
pydantic
dotenv